# Port to django-pipeline 1.6.7