--- a/requirements.txt
+++ b/requirements.txt
@@ -19,7 +19,7 @@ diff-match-patch==20121119
 django-auth-ldap==1.2.6
 django-bitfield==1.8.0
 git+https://github.com/rwbarton/django-guardian.git@caf9f0c8c035feb3dff5542fb042dd13126cdd69
-django-pipeline==1.2.2
+django-pipeline==1.6.7
 docopt==0.4.0
 enum34==1.0.4
 fonttools==3.0