Now Zulip is compatible with django-pipeline 1.6.7. It is no more compatible with the older version of django-pipeline (1.2.2). This will also help in adding Python 3 support.