@sharmaeklavya2 2016-03-17
How to test it:

1. Run `tools/minify-js`.
2. Set `DEBUG = False` in zproject/settings.py.
3. Collect static files - `mkdir -p prod-static/serve` followed by `python manage.py collectstatic`.
 
Now run the development server (`tools/run-dev.py --interface=''`).