@timabbott 2016-03-17
Have you also tested the `tools/build-release-tarball` codepath?  You
should be able to compare the resulting static assets to see if they're the
same before and after your change.


        -Tim Abbott

On Thu, Mar 17, 2016 at 9:45 AM, Eklavya Sharma <notifications@github.com>
wrote:

> How to test it:
>
>    1. Run tools/minify-js.
>    2. Set DEBUG = False in zproject/settings.py.
>    3. Collect static files - mkdir -p prod-static/serve followed by python
>    manage.py collectstatic.
>
> Now run the development server (tools/run-dev.py --interface='').
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly or view it on GitHub
> <https://github.com/zulip/zulip/pull/545#issuecomment-197969264>
>