@sharmaeklavya2 2016-03-17
I couldn't figure out what you mean by that. Can you please help me out? What exactly should I do?

I tried running tools/build-release-tarball, but it told me to give `ZULIP_VERSION` as a command line argument. What should I pass as `ZULIP_VERSION`?

This will probably create a tarball (as the script's name suggests). What should I do with that tarball?