@timabbott 2016-03-17
Pass in any string you like as ZULIP_VERSION, e.g. "test".  It generates
the tarball based on the code on HEAD.

​It will create a tarball; what you should do to validate this is create
the tarball from git commits before and after your changes, unpack the two
tarballs, and compare the files inside; the actual filenames will be
different since the unique VFL keys will be different, but you may be able
to get of info out of diffs (maybe with something like wdiff) the output
static asset files, which mostly are in the prod-static/ directory of the
tarballs.