@timabbott 2016-03-18
Great, thanks for the contribution @sharmaeklavya2 !  This looks pretty good; I posted one comment.

Just to confirm, the JS files were the same?  

As a note for myself, we need to upgrade django-pipeline in the Zulip apt repository as well as part of merging this.