@sharmaeklavya2 2016-03-18
Yes the JS files were the same.