@timabbott 2016-03-19
Hmm, I think that one is probably OK to have as part of this commit, but I'd mention in the commit message that you had to change the mechanism for removing unminified CSS because of this issue.