@timabbott 2016-03-31
Hey @sharmaeklavya2!  I just built a django-pipeline package for the apt repository in preparation for merging this, and tested it out.  Empirically, it seems that the loading times for pages that use the template system are slower with the new version (I don't have a solid reproduction, but with 1.2.2  `/integrations` takes 100-500ms on my dev environment and with 1.6.7 it seems to be 185-900ms).  I'm not 100% confident this is a real effect, but I think it's enough to be worth investigating.  It could make sense to try running the two versions under cProfile to compare the results.  

I think it should be possible to use either a stripped-down version of zilencer/management/commands/profile_request.py or perhaps better, `test-backend --profile` using a test that just visits one of these external URLs to get a clear profile data set for what the backend is doing for what fraction of the time and get a strong sense of whether this is in fact slower.  If it is, there's a few approaches we could take.  I'm also going to try testing this in a more controlled environment to get more data.