@timabbott 2016-03-31
OK I've confirmed the new version is seems quite likely slower at building static assets (which happens at runtime in development but not in prod because in prod we cache the templates; it's only slightly slower at runtime in prod).  Here's some stats:

* `build-release-tarball` runtime (builds static assets): 2m44s => 3m32s
* prod-style test machine `/integrations`: Avg goes from 7ms =>9ms
* development `/integrations`: Avg goes from 200ms=>400ms.

@sharmaeklavya2 are you up for investigating this performance mismatch?  It appears this does not heavily affect the production runtime performance, but slowing down development feedback cycles is important too...