@sharmaeklavya2 2016-04-02
I have never profiled programs in the past, so I'll have to first learn how to do that.

@timabbott, what do the stats on `/integrations` mean? What is it whose time you have measured there?