@timabbott 2016-04-02
So I'm doing those measurements by just watching the `run-dev.py` console in development and /var/log/zulip/server.og in production -- we log the runtime for every request.  The code that does this logging is in `write_log_line` and its helpers in zerver/middleware.py.

/integrations is way faster in prod because of this block in zproject/settings.py:
```
if PRODUCTION:                                                                                                          
    # Template caching is a significant performance win in production.                                                  
    TEMPLATE_LOADERS = (                                                                                                
        ('django.template.loaders.cached.Loader',                                                                       
         TEMPLATE_LOADERS),                                                                                             
        )                                                                                                               
```
So most of the processing it only does once per template rather than once per request.  We can't use that in development right now, because the way that caching works it doesn't update the cache when a user changes the template content.  I suppose one potential solution to this overall problem might be to add some mechanism for invalidating that cache whenever changes are made in the repository (we certainly already have autoreload code listening for changes...); I opened #627 based on this idea.

Another approach would be to try using a faster templating engine (see #620); depending on why this new version is slower, that might address the issue here.