@timabbott 2016-05-09
Just wanted to mention that @umairwaheed opened https://github.com/jazzband/django-pipeline/issues/567 to get upstream aware of the performance issue.