@timabbott 2016-06-28
Now that #717 is resolved, I think it makes sense to pick this up again.