@sharmaeklavya2 2016-06-28
I was keeping this pull request up to date by frequently rebasing on master, in the hope that the performance issues will eventually be resolved and this will be merged. But I could no longer do that after we switched to Jinja2. This pull request is now far behind master.