@timabbott 2016-06-28
Yeah.  It's possible that now that we're on Jinja2, the performance issues are less severe.  @umairwaheed, I know you've now done a lot of investigation into our static build process as part of doing the Jinja2 migration; do you want to take over finishing this?  I think you might be better equipped to delete the hacks that we no longer need when moving to the new version...