@umairwaheed 2016-07-04
@timabbott , @sharmaeklavya2 done in https://github.com/umairwaheed/zulip/tree/pipeline-upgrade