@umairwaheed 2016-07-05
@timabbott, on slowness of Django-Pipeline, the upstream has identified an issue in case of `DEBUG=True` https://github.com/jazzband/django-pipeline/issues/572.