@timabbott 2016-07-06
OK.  That doesn't explain the `update-prod-static` issue, though...